removed redundant create_borrow_request endpoint and moved items endpoints to seperate routes file
