+++ backend/routes/items.py
@@ -1,0 +1,105 @@
+
+from flask import Blueprint, request, jsonify, session
+from backend.models import Item, User
+from backend.extensions import db
+from datetime import datetime
+
+items_bp = Blueprint('items', __name__)
+
+@items_bp.route('/items', methods=['POST', 'GET'])
+def create_item():
+    # GET
+    if request.method == 'GET':
+        items = Item.query.filter_by(available=True).order_by(Item.created_at.desc()).all()
+        return jsonify([{
+            "id": item.id,
+            "item_name": item.item_name,
+            "description": item.description,
+            "category": item.category,
+            "size": item.size,
+            "brand": item.brand,
+            "condition": item.condition,
+            "price_per_day": item.price_per_day,
+            "images": item.images,
+            "available": item.available,
+            "created_at": item.created_at.isoformat()
+        } for item in items])
+
+    # POST
+    user_id = session.get('user_id')
+    if not user_id:
+        return jsonify({"error": "You must be logged in to list an item"}), 401
+
+    data = request.get_json(silent=True)
+    if not data:
+        return jsonify({"error": "No data provided"}), 400
+
+    try:
+        new_item = Item(
+            owner_id=user_id,
+            item_name=data.get('item_name'),
+            description=data.get('description'),
+            category=data.get('category'),
+            size=data.get('size'),
+            brand=data.get('brand'),
+            condition=data.get('condition'),
+            price_per_day=int(data.get('price_per_day', 0)),
+            link=data.get('link'),
+            images=data.get('images', []),
+            available=True
+        )
+        db.session.add(new_item)
+        db.session.commit()
+        return jsonify({"message": "Item listed successfully!", "id": new_item.id}), 201
+
+    except Exception as e:
+        db.session.rollback()
+        return jsonify({"error": "Could not create item"}), 400
+
+
+@items_bp.route('/items/<int:item_id>', methods=['GET'])
+def get_item(item_id):
+    item = Item.query.get(item_id)
+    if not item:
+        return jsonify({'error': 'Not found'}), 404
+
+    owner = User.query.get(item.owner_id)
+
+    return jsonify({
+        'id': item.id,
+        'item_name': item.item_name,
+        'description': item.description,
+        'category': item.category,
+        'size': item.size,
+        'images': item.images,
+        'available': item.available,
+        'brand': item.brand,
+        'condition': item.condition,
+        'price_per_day': item.price_per_day,
+        'owner_id': item.owner_id,
+        'owner_name': f"{owner.firstName} {owner.lastName}",
+    }), 200
+
+
+@items_bp.route('/my-items', methods=['GET'])
+def get_my_items():
+    user_id = session.get('user_id')
+    if not user_id:
+        return jsonify({"error": "You must be logged in"}), 401
+
+    items = Item.query.filter_by(owner_id=user_id).order_by(Item.created_at.desc()).all()
+
+    return jsonify([{
+        "id": item.id,
+        "owner_id": item.owner_id,
+        "item_name": item.item_name,
+        "description": item.description,
+        "category": item.category,
+        "size": item.size,
+        "brand": item.brand,
+        "condition": item.condition,
+        "price_per_day": item.price_per_day,
+        "images": item.images,
+        "available": item.available,
+        "created_at": item.created_at.isoformat()
+    } for item in items])
